Separated Message MessageType and MessageException
