+++ src/main/java/org/qortal/network/message/MessageException.java
@@ -1,0 +1,18 @@
+package org.qortal.network.message;
+
+public class MessageException extends Exception {
+    public MessageException() {
+    }
+
+    public MessageException(String message) {
+        super(message);
+    }
+
+    public MessageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public MessageException(Throwable cause) {
+        super(cause);
+    }
+}
